Added face detection model
This model is deployed in AWS. This ensures that the user uploads at least one image with his face in his/her profile. This script classifies the image as Accepted or Not Accepted based on whether it can detect a face in the image or not. A valid photo necessitates at least one prominent face in the image.